extracts createBaseColumns and reorganizes
