+++ src/components/network-modification-table/renderers/index.ts
@@ -1,0 +1,13 @@
+/**
+ import NetworkModificationsTable from './network-modification-table/network-modifications-table';
+ * Copyright (c) 2026, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+export * from './drag-handle-cell';
+export * from './name-cell';
+export * from './network-modification-node-editor-name-header';
+export * from './select-cell';
+export * from './select-header-cell';
